Fix "make it your own" command pipeline
Because the Git repository is deleted before creating a new repository
with no history (just to be sure there is no traceback to the original
author), it is necessary to make a commit before pushing to the remote
repository, or else an error occurs.

This commit also adds the note to create a GitHub empty repository to
be able to push the modified repository.